requirements.txt
discord.py >= 2.0.1
 pillow >= 8.3.0
 requests >= 2.23.0 
 tortoise-orm >= 0.17.6
 markovify >= 0.9.3
 pynacl >= 1.3.0 
 youtube-dl >= 2020.5.8 
 gtts >= 2.1.1
 git+https://github.com/Rapptz/discord-ext-menus
 humanize >= 2.6.0
 psutil >= 5.9.0
